// src/tools/utils/responsive/enums.ts
/* eslint-disable no-unused-vars */
export enum Breakpoint {
  mobile = "mobile",
  tabletPortrait = "tabletPortrait",
  tabletLandscape = "tabletLandscape",
  desktopSmall = "desktopSmall",
  desktopLarge = "desktopSmall",
}

export type BreakpointLiteral =
  | "mobile"
  | "tabletPortait"
  | "tabletLandscape"
  | "desktopSmall"
  | "desktopLarge";